Rename KuzzleRequest to RequestPayload
