+++ src/controllers/Base.ts
@@ -1,4 +1,4 @@
-import { KuzzleRequest, JSONObject } from '../utils/interfaces';
+import { RequestPayload, JSONObject } from '../types';
 
 export class BaseController {
   private _name: string;
@@ -34,7 +34,7 @@ export class BaseController {
    * @param request
    * @param options
   */
-  query (request: KuzzleRequest = {}, options: any = {}): Promise<JSONObject> {
+  query (request: RequestPayload = {}, options: any = {}): Promise<JSONObject> {
     request.controller = request.controller || this.name;
 
     return this._kuzzle.query(request, options);
