+++ src/types/ApiKey.ts
@@ -1,0 +1,34 @@
+/**
+ * ApiKey
+ */
+export type ApiKey = {
+  /**
+   * ApiKey unique ID
+   */
+  _id: string;
+  /**
+   * ApiKey content
+   */
+  _source: {
+    /**
+     * User kuid
+     */
+    userId: string;
+    /**
+     * Expiration date in Epoch-millis format (-1 if the token never expires)
+     */
+    expiresAt: number;
+    /**
+     * Original TTL in ms
+     */
+    ttl: number;
+    /**
+     * API key description
+     */
+    description: string;
+    /**
+     * Authentication token associated with this API key
+     */
+    token: string;
+  }
+}
