+++ src/types/Document.ts
@@ -1,0 +1,72 @@
+import { JSONObject } from './JSONObject';
+
+/**
+ * Kuzzle metadata
+ * @see https://docs.kuzzle.io/core/2/guides/essentials/document-metadata/
+ */
+export interface DocumentMetadata {
+  _kuzzle_info: {
+    /**
+     * Kuid of the user who created the document
+     */
+    author: string,
+    /**
+     * Creation date in micro-timestamp
+     */
+    createdAt: number,
+    /**
+     * Kuid of the user who last updated the document
+     */
+    updater: string | null,
+    /**
+     * Update date in micro-timestamp
+     */
+    updatedAt: number | null
+  }
+}
+
+/**
+ * Represents the `_source` property of the document
+ */
+export interface DocumentContent extends DocumentMetadata {
+  [key: string]: JSONObject | any,
+}
+
+/**
+ * Kuzzle document
+ *
+ * @property _id
+ * @property _version
+ * @property _source
+ */
+export class Document {
+  /**
+   * Document unique ID
+   */
+  _id: string;
+  /**
+   * Document Version (generated by Elasticsearch)
+   */
+  _version?: number;
+  /**
+   * Document Content
+   */
+  _source: DocumentContent;
+
+  [key: string]: JSONObject | any;
+}
+
+/**
+ * Document retrieved from a search
+ *
+ * @property _id
+ * @property _version
+ * @property _source
+ * @property _score
+ */
+export interface DocumentHit extends Document {
+  /**
+   * Relevance score
+   */
+  _score: number;
+}
