+++ src/types/JSONObject.ts
@@ -1,0 +1,6 @@
+/**
+ * An interface representing an object with string key and any value
+ */
+export interface JSONObject {
+  [key: string]: JSONObject | any
+}
